Fixed subtle bug with comparison of return value with sizeof
git-svn-id: svn+ssh://wicontroller4/home/svn/wirover_kernel/trunk@301 ba602a36-4df9-46f2-848b-6875ca4683ef

+++ wigateway/bandwidth.c
@@ -400,7 +400,7 @@ static int recv_burst_udp(struct bw_client_info *client, struct bw_stats *stats,
 
         result = recvfrom_timeout(sockfd, buffer, buffer_len, 0,
                 (struct sockaddr *)&sender_addr, &sender_addr_len, &timeout);
-        if(result >= sizeof(struct bw_hdr)) {
+        if(result >= (int)sizeof(struct bw_hdr)) {
             struct bw_hdr *bw_hdr = (struct bw_hdr *)buffer;
 
             // TODO: Verify sender address matches server
